update plugin all-in-one-seo-pack to version 2.4.3.1

+++ wp-content/plugins/all-in-one-seo-pack/all_in_one_seo_pack.php
@@ -1,22 +1,21 @@
 <?php
 /*
 Plugin Name: All In One SEO Pack
-Plugin URI: http://semperfiwebdesign.com
-Description: Out-of-the-box SEO for your WordPress blog. Features like XML Sitemaps, SEO for custom post types, SEO for blogs or business sites, SEO for ecommerce sites, and much more. Almost 30 million downloads since 2007.
-Version: 2.3.8
+Plugin URI: https://semperplugins.com/all-in-one-seo-pack-pro-version/
+Description: Out-of-the-box SEO for your WordPress blog. Features like XML Sitemaps, SEO for custom post types, SEO for blogs or business sites, SEO for ecommerce sites, and much more. More than 30 million downloads since 2007.
+Version: 2.4.3.1
 Author: Michael Torbert
-Author URI: http://michaeltorbert.com
+Author URI: https://semperplugins.com/all-in-one-seo-pack-pro-version/
 Text Domain: all-in-one-seo-pack
 Domain Path: /i18n/
 */
 
 /*
-Copyright (C) 2007-2016 Michael Torbert, https://semperfiwebdesign.com
+Copyright (C) 2007-2017 Michael Torbert, https://semperfiwebdesign.com
 
 This program is free software; you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
-the Free Software Foundation; either version 3 of the License, or
-(at your option) any later version.
+the Free Software Foundation; version 2 of the License.
 
 This program is distributed in the hope that it will be useful,
 but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -32,14 +31,14 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
  * The original WordPress SEO plugin.
  *
  * @package All-in-One-SEO-Pack
- * @version 2.3.8
+ * @version 2.4.3.1
  */
 
 if ( ! defined( 'AIOSEOPPRO' ) ) {
 	define( 'AIOSEOPPRO', false );
 }
 if ( ! defined( 'AIOSEOP_VERSION' ) ) {
-	define( 'AIOSEOP_VERSION', '2.3.8' );
+	define( 'AIOSEOP_VERSION', '2.4.3.1' );
 }
 global $aioseop_plugin_name;
 $aioseop_plugin_name = 'All in One SEO Pack';
@@ -54,7 +53,7 @@ if ( ! defined( 'ABSPATH' ) ) {
 
 if ( AIOSEOPPRO ) {
 
-	add_action( 'admin_init', 'disable_all_in_one_free', 1 );
+	add_action( 'admin_head', 'disable_all_in_one_free', 1 );
 
 }
 
@@ -78,7 +77,6 @@ if ( ! defined( 'AIOSEOP_PLUGIN_NAME' ) ) {
 	define( 'AIOSEOP_PLUGIN_NAME', $aioseop_plugin_name );
 }
 
-// Do we need this? register_activation_hook(__FILE__,'aioseop_activate_pl');.
 if ( ! defined( 'AIOSEOP_PLUGIN_DIR' ) ) {
 	define( 'AIOSEOP_PLUGIN_DIR', plugin_dir_path( __FILE__ ) );
 } elseif ( AIOSEOP_PLUGIN_DIR !== plugin_dir_path( __FILE__ ) ) {
@@ -217,8 +215,12 @@ if ( class_exists( 'All_in_One_SEO_Pack' ) ) {
 if ( AIOSEOPPRO ) {
 
 	require( AIOSEOP_PLUGIN_DIR . 'pro/sfwd_update_checker.php' );
+	$aiosp_update_url = 'https://semperplugins.com/upgrade_plugins.php';
+	if( defined( 'AIOSEOP_UPDATE_URL' ) ) {
+		$aiosp_update_url = AIOSEOP_UPDATE_URL;
+	}
 	$aioseop_update_checker = new SFWD_Update_Checker(
-		'http://semperplugins.com/upgrade_plugins.php',
+		$aiosp_update_url,
 		__FILE__,
 		'aioseop'
 	);
@@ -230,8 +232,8 @@ if ( AIOSEOPPRO ) {
 	} else {
 		$aioseop_update_checker->license_key = '';
 	}
-	$aioseop_update_checker->options_page = 'all-in-one-seo-pack-pro/aioseop_class.php';
-	$aioseop_update_checker->renewal_page = 'http://semperplugins.com/all-in-one-seo-pack-pro-support-updates-renewal/';
+	$aioseop_update_checker->options_page = AIOSEOP_PLUGIN_DIRNAME . "/aioseop_class.php";
+	$aioseop_update_checker->renewal_page = 'https://semperplugins.com/all-in-one-seo-pack-pro-version/';
 
 	$aioseop_update_checker->addQueryArgFilter( array( $aioseop_update_checker, 'add_secret_key' ) );
 }
@@ -247,6 +249,11 @@ if ( ! function_exists( 'aioseop_activate' ) ) {
 			global $aioseop_update_checker;
 		}
 		$aiosp_activation = true;
+
+		// These checks might be duplicated in the function being called.
+		if( ! is_network_admin() || !isset( $_GET['activate-multi'] ) ) {
+			set_transient( '_aioseop_activation_redirect', true, 30 ); // Sets 30 second transient for welcome screen redirect on activation.
+			}
 
 		delete_user_meta( get_current_user_id(), 'aioseop_yst_detected_notice_dismissed' );
 
@@ -270,23 +277,16 @@ if ( ! function_exists( 'aiosp_plugin_row_meta' ) ) {
 	 */
 	function aiosp_plugin_row_meta( $actions, $plugin_file ) {
 
-		if ( ! AIOSEOPPRO ) {
 
 			$action_links = array(
-				'donatelink' => array(
-					'label' => __( 'Donate', 'all-in-one-seo-pack' ),
-					'url'   => 'https://www.paypal.com/cgi-bin/webscr?cmd=_donations&business=mrtorbert%40gmail%2ecom&item_name=All%20In%20One%20SEO%20Pack&item_number=Support%20Open%20Source&no_shipping=0&no_note=1&tax=0&currency_code=USD&lc=US&bn=PP%2dDonationsBF&charset=UTF%2d8',
-				),
-				'amazon'     => array(
-					'label' => __( 'Amazon Wishlist', 'all-in-one-seo-pack' ),
-					'url'   => 'https://www.amazon.com/wishlist/1NFQ133FNCOOA/ref=wl_web',
-				),
+
+				'settings' => array(
+					'label' => __( 'Feature Request/Bug Report', 'all-in-one-seo-pack' ),
+					'url'   => 'https://github.com/semperfiwebdesign/all-in-one-seo-pack/issues/new' )
 
 			);
 
-		} else {
-			$action_links = '';
-		}
+
 
 		return aiosp_action_links( $actions, $plugin_file, $action_links, 'after' );
 	}
@@ -304,6 +304,9 @@ if ( ! function_exists( 'aiosp_add_action_links' ) ) {
 	 * @return array
 	 */
 	function aiosp_add_action_links( $actions, $plugin_file ) {
+		if ( ! is_array( $actions ) ) {
+			return $actions;
+		}
 
 		$aioseop_plugin_dirname = AIOSEOP_PLUGIN_DIRNAME;
 		$action_links           = array();
@@ -315,12 +318,12 @@ if ( ! function_exists( 'aiosp_add_action_links' ) ) {
 
 			'forum' => array(
 				'label' => __( 'Support Forum', 'all-in-one-seo-pack' ),
-				'url'   => 'http://semperplugins.com/support/',
+				'url'   => 'https://semperplugins.com/support/',
 			),
 
 			'docs' => array(
 				'label' => __( 'Documentation', 'all-in-one-seo-pack' ),
-				'url'   => 'http://semperplugins.com/documentation/',
+				'url'   => 'https://semperplugins.com/documentation/',
 			),
 
 		);
@@ -331,7 +334,7 @@ if ( ! function_exists( 'aiosp_add_action_links' ) ) {
 			$action_links['proupgrade'] =
 				array(
 					'label' => __( 'Upgrade to Pro', 'all-in-one-seo-pack' ),
-					'url'   => 'http://semperplugins.com/plugins/all-in-one-seo-pack-pro-version/?loc=plugins',
+					'url'   => 'https://semperplugins.com/plugins/all-in-one-seo-pack-pro-version/?loc=plugins',
 
 				);
 		}
@@ -369,8 +372,13 @@ if ( ! function_exists( 'aiosp_action_links' ) ) {
 	}
 }
 
-
 if ( ! function_exists( 'aioseop_init_class' ) ) {
+	/**
+	 * Inits All-in-One-Seo plugin class.
+	 *
+	 * @since ?? // When was this added?
+	 * @since 2.3.12.3 Loads third party compatibility class.
+	 */
 	function aioseop_init_class() {
 		global $aiosp;
 		load_plugin_textdomain( 'all-in-one-seo-pack', false, dirname( plugin_basename( __FILE__ ) ) . '/i18n/' );
@@ -384,13 +392,18 @@ if ( ! function_exists( 'aioseop_init_class' ) ) {
 		require_once( AIOSEOP_PLUGIN_DIR . 'admin/meta_import.php' );
 		require_once( AIOSEOP_PLUGIN_DIR . 'inc/translations.php' );
 		require_once( AIOSEOP_PLUGIN_DIR . 'public/opengraph.php' );
+		require_once( AIOSEOP_PLUGIN_DIR . 'inc/compatability/abstract/aiosep_compatible.php');
 		require_once( AIOSEOP_PLUGIN_DIR . 'inc/compatability/compat-init.php');
 		require_once( AIOSEOP_PLUGIN_DIR . 'public/front.php' );
+		require_once( AIOSEOP_PLUGIN_DIR . 'public/google-analytics.php' );
+		require_once( AIOSEOP_PLUGIN_DIR . 'admin/display/welcome.php' );
+		require_once( AIOSEOP_PLUGIN_DIR . 'admin/display/dashboard_widget.php' );
+		require_once( AIOSEOP_PLUGIN_DIR . 'admin/display/menu.php' );
+
+		$aioseop_welcome = new aioseop_welcome(); // TODO move this to updates file.
 
 		if ( AIOSEOPPRO ) {
-			require_once( AIOSEOP_PLUGIN_DIR . 'pro/functions_general.php' );
-			require_once( AIOSEOP_PLUGIN_DIR . 'pro/functions_class.php' );
-			require_once( AIOSEOP_PLUGIN_DIR . 'pro/aioseop_pro_updates_class.php' );
+			require_once( AIOSEOP_PLUGIN_DIR . 'pro/class-aio-pro-init.php' ); // Loads pro files and other pro init stuff.
 		}
 		aiosp_seometa_import(); // call importer functions... this should be moved somewhere better
 
@@ -402,6 +415,8 @@ if ( ! function_exists( 'aioseop_init_class' ) ) {
 			$aioseop_pro_updates = new AIOSEOP_Pro_Updates();
 			add_action( 'admin_init', array( $aioseop_pro_updates, 'version_updates' ), 12 );
 		}
+
+		add_action( 'admin_init', 'aioseop_welcome' );
 
 		if ( aioseop_option_isset( 'aiosp_unprotect_meta' ) ) {
 			add_filter( 'is_protected_meta', 'aioseop_unprotect_meta', 10, 3 );
@@ -423,6 +438,19 @@ if ( ! function_exists( 'aioseop_init_class' ) ) {
 	}
 }
 
+
+
+if ( ! function_exists( 'aioseop_welcome' ) ){
+	function aioseop_welcome(){
+		if( get_transient( '_aioseop_activation_redirect') ){
+			$aioseop_welcome = new aioseop_welcome();
+			delete_transient( '_aioseop_activation_redirect' );
+			$aioseop_welcome->init( TRUE );
+		}
+
+	}
+}
+
 add_action( 'init', 'aioseop_load_modules', 1 );
 //add_action( 'after_setup_theme', 'aioseop_load_modules' );
 
@@ -439,6 +467,7 @@ if ( is_admin() ) {
 	add_action( 'wp_ajax_aioseo_dismiss_yst_notice', 'aioseop_update_yst_detected_notice' );
 	add_action( 'wp_ajax_aioseo_dismiss_visibility_notice', 'aioseop_update_user_visibilitynotice' );
 	add_action( 'wp_ajax_aioseo_dismiss_woo_upgrade_notice', 'aioseop_woo_upgrade_notice_dismissed' );
+	add_action( 'wp_ajax_aioseo_dismiss_sitemap_max_url_notice', 'aioseop_sitemap_max_url_notice_dismissed' );
 	if ( AIOSEOPPRO ) {
 		add_action( 'wp_ajax_aioseop_ajax_update_oembed', 'aioseop_ajax_update_oembed' );
 	}
